--- QMMHarmonyShimmer/Harmony/Priority.cs
@@ -1,15 +1,0 @@
-﻿namespace Harmony
-{
-	public static class Priority
-	{
-		public const int Last = 0;
-		public const int VeryLow = 100;
-		public const int Low = 200;
-		public const int LowerThanNormal = 300;
-		public const int Normal = 400;
-		public const int HigherThanNormal = 500;
-		public const int High = 600;
-		public const int VeryHigh = 700;
-		public const int First = 800;
-	}
-}
