--- QMMHarmonyShimmer/Patches/AccessToolsDeclaredFieldShim.cs
@@ -1,18 +1,0 @@
-﻿using System;
-using System.Reflection;
-using HarmonyLib;
-
-namespace QModManager.QMMHarmonyShimmer.Patches
-{
-    [HarmonyPatch(typeof(AccessTools), nameof(AccessTools.DeclaredField), typeof(Type), typeof(string))]
-    internal static class AccessToolsDeclaredFieldShim
-    {
-        private static void Postfix(ref FieldInfo __result, Type type, string name)
-        {
-            if (__result == null)
-            {
-                __result = AccessTools.Field(type, name);
-            }
-        }
-    }
-}
