--- QMMHarmonyShimmer/QModManager.QMMHarmonyShimmer.csproj
@@ -1,76 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<Project ToolsVersion="15.0" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
-  <Import Project="$(MSBuildExtensionsPath)\$(MSBuildToolsVersion)\Microsoft.Common.props" Condition="Exists('$(MSBuildExtensionsPath)\$(MSBuildToolsVersion)\Microsoft.Common.props')" />
-  <PropertyGroup>
-    <Configuration Condition=" '$(Configuration)' == '' ">Debug</Configuration>
-    <Platform Condition=" '$(Platform)' == '' ">AnyCPU</Platform>
-    <ProjectGuid>{F308F15A-2767-4B4E-BAA4-D9B95D99F8E5}</ProjectGuid>
-    <OutputType>Library</OutputType>
-    <AppDesignerFolder>Properties</AppDesignerFolder>
-    <RootNamespace>QModManager.QMMHarmonyShimmer</RootNamespace>
-    <AssemblyName>QModManager.QMMHarmonyShimmer</AssemblyName>
-    <TargetFrameworkVersion>v4.0</TargetFrameworkVersion>
-    <FileAlignment>512</FileAlignment>
-    <Deterministic>true</Deterministic>
-    <TargetFrameworkProfile />
-  </PropertyGroup>
-  <PropertyGroup Condition=" '$(Configuration)|$(Platform)' == 'Debug|AnyCPU' ">
-    <DebugSymbols>true</DebugSymbols>
-    <DebugType>full</DebugType>
-    <Optimize>false</Optimize>
-    <OutputPath>..\Build\</OutputPath>
-    <DefineConstants>DEBUG;TRACE</DefineConstants>
-    <ErrorReport>prompt</ErrorReport>
-    <WarningLevel>4</WarningLevel>
-    <AllowUnsafeBlocks>true</AllowUnsafeBlocks>
-  </PropertyGroup>
-  <PropertyGroup Condition=" '$(Configuration)|$(Platform)' == 'Release|AnyCPU' ">
-    <DebugType>pdbonly</DebugType>
-    <Optimize>true</Optimize>
-    <OutputPath>..\Build\</OutputPath>
-    <DefineConstants>TRACE</DefineConstants>
-    <ErrorReport>prompt</ErrorReport>
-    <WarningLevel>4</WarningLevel>
-    <AllowUnsafeBlocks>true</AllowUnsafeBlocks>
-  </PropertyGroup>
-  <ItemGroup>
-    <Reference Include="0Harmony">
-      <HintPath>..\Dependencies\BepInEx\BepInEx\core\0Harmony.dll</HintPath>
-      <Private>False</Private>
-    </Reference>
-    <Reference Include="BepInEx">
-      <HintPath>..\Dependencies\BepInEx\BepInEx\core\BepInEx.dll</HintPath>
-      <Private>False</Private>
-    </Reference>
-    <Reference Include="Mono.Cecil">
-      <HintPath>..\Dependencies\BepInEx\BepInEx\core\Mono.Cecil.dll</HintPath>
-      <Private>False</Private>
-    </Reference>
-    <Reference Include="System" />
-  </ItemGroup>
-  <ItemGroup>
-    <Compile Include="Harmony\AccessCache.cs" />
-    <Compile Include="Harmony\AccessTools.cs" />
-    <Compile Include="Harmony\Attributes.cs" />
-    <Compile Include="Harmony\Extensions.cs" />
-    <Compile Include="Harmony\FileLog.cs" />
-    <Compile Include="Harmony\HarmonyInstance.cs" />
-    <Compile Include="Harmony\HarmonyMethod.cs" />
-    <Compile Include="Harmony\HarmonySharedState.cs" />
-    <Compile Include="Harmony\MethodInvoker.cs" />
-    <Compile Include="Harmony\Patch.cs" />
-    <Compile Include="Harmony\PatchFunctions.cs" />
-    <Compile Include="Harmony\PatchProcessor.cs" />
-    <Compile Include="Harmony\PatchTools.cs" />
-    <Compile Include="Harmony\Priority.cs" />
-    <Compile Include="Patches\AccessToolsDeclaredFieldShim.cs" />
-    <Compile Include="Patches\TypeDefinitionIsDefinitionPatch.cs" />
-    <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="Harmony\ShimUtil.cs" />
-    <Compile Include="Harmony\SymbolExtensions.cs" />
-    <Compile Include="Harmony\Traverse.cs" />
-    <Compile Include="QMMHarmonyShimmer.cs" />
-  </ItemGroup>
-  <ItemGroup />
-  <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
-</Project>
