--- QMMHarmonyShimmer/README.md
@@ -1,3 +1,0 @@
-# QMMHarmonyShimmer
-
-A patcher for [BepInEx](https://bepinex.github.io/bepinex_docs/) to shim Harmony references in QMods pre-Harmony 2, largely inspired by and based on the work of [IPALoaderX](https://github.com/BepInEx/IPALoaderX)'s [BepInEx.HarmonyShimmer](https://github.com/BepInEx/IPALoaderX/tree/master/BepInEx.IPAHarmonyShimmer).
