+++ QModManager.sln
@@ -20,8 +20,6 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Executable", "Executable\Ex
 		{CA99F076-7459-4682-A41D-2850AEE28566} = {CA99F076-7459-4682-A41D-2850AEE28566}
 	EndProjectSection
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "QModManager.QMMHarmonyShimmer", "QMMHarmonyShimmer\QModManager.QMMHarmonyShimmer.csproj", "{F308F15A-2767-4B4E-BAA4-D9B95D99F8E5}"
-EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "QModManager.QMMLoader", "QMMLoader\QModManager.QMMLoader.csproj", "{21D282E0-6B3E-4776-AC86-3089F0655893}"
 	ProjectSection(ProjectDependencies) = postProject
 		{EA496DDF-D775-4E17-9EAE-C570C5F50701} = {EA496DDF-D775-4E17-9EAE-C570C5F50701}
@@ -91,18 +89,6 @@ Global
 		{E00B7FE8-0F1D-4AE6-9E47-4BFD81537F14}.Release|x64.Build.0 = Release|x64
 		{E00B7FE8-0F1D-4AE6-9E47-4BFD81537F14}.Release|x86.ActiveCfg = Release|x86
 		{E00B7FE8-0F1D-4AE6-9E47-4BFD81537F14}.Release|x86.Build.0 = Release|x86
-		{F308F15A-2767-4B4E-BAA4-D9B95D99F8E5}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{F308F15A-2767-4B4E-BAA4-D9B95D99F8E5}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{F308F15A-2767-4B4E-BAA4-D9B95D99F8E5}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{F308F15A-2767-4B4E-BAA4-D9B95D99F8E5}.Debug|x64.Build.0 = Debug|Any CPU
-		{F308F15A-2767-4B4E-BAA4-D9B95D99F8E5}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{F308F15A-2767-4B4E-BAA4-D9B95D99F8E5}.Debug|x86.Build.0 = Debug|Any CPU
-		{F308F15A-2767-4B4E-BAA4-D9B95D99F8E5}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{F308F15A-2767-4B4E-BAA4-D9B95D99F8E5}.Release|Any CPU.Build.0 = Release|Any CPU
-		{F308F15A-2767-4B4E-BAA4-D9B95D99F8E5}.Release|x64.ActiveCfg = Release|Any CPU
-		{F308F15A-2767-4B4E-BAA4-D9B95D99F8E5}.Release|x64.Build.0 = Release|Any CPU
-		{F308F15A-2767-4B4E-BAA4-D9B95D99F8E5}.Release|x86.ActiveCfg = Release|Any CPU
-		{F308F15A-2767-4B4E-BAA4-D9B95D99F8E5}.Release|x86.Build.0 = Release|Any CPU
 		{21D282E0-6B3E-4776-AC86-3089F0655893}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{21D282E0-6B3E-4776-AC86-3089F0655893}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{21D282E0-6B3E-4776-AC86-3089F0655893}.Debug|x64.ActiveCfg = Debug|Any CPU
