+++ QModManager/Patching/ManifestValidator.cs
@@ -241,7 +241,7 @@
             }
             catch (ReflectionTypeLoadException rtle)
             {
-                Logger.Debug($"Unable to load types for '{qMod.Id}': " + rtle.Message);
+                Logger.Debug($"Unable to load types for '{qMod.Id}': \nInnerException: \n" + rtle.InnerException + "\n LoaderExceptions:\n" +string.Join("/n", rtle.LoaderExceptions.ToList()));
                 qMod.Status = ModStatus.MissingDependency;
             }
